Renamed `Store` to `FileStore` to better describe its purpose.
