+++ lib/provider/local/index.js
@@ -6,7 +6,7 @@ const REQUIRED_HEADERS = ['authorization', 'date', 'digest', 'host'];
 const DEFAULT_SKEW = 5000;
 const AUTHN_PROTOCOL = 'Rapid7-HMAC-V1-SHA256';
 
-const Store = exports.Store = require('./store');
+const FileStore = exports.FileStore = require('./file_store');
 
 /**
  * Validate a request's authentication headers against a local database of keys
@@ -117,7 +117,7 @@ exports.authn = function(options) {
   const algorithm = options.algorithm;
   const skew = Number(options.skew);
 
-  const db = new Store(options.db);
+  const db = new FileStore(options.db);
 
   db.on('error', (err) => Log.error(err));
